move files to driveCommands folder
