+++ src/main/java/com/spartronics4915/frc2025/commands/DriveCommands/SwerveTeleopCommand.java
@@ -1,4 +1,4 @@
-package com.spartronics4915.frc2025.commands;
+package com.spartronics4915.frc2025.commands.DriveCommands;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.kinematics.ChassisSpeeds;
